Making very clear that this isn't in the final release of the SDK
